backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic>=0.40.0
numpy==1.26.2
scipy==1.11.4
pydantic==2.5.0
python-dotenv==1.0.0
